Feature/81 (#83)
* apply splash page and login page

* refactor UI components and improve pin memo layout

* refactor follow item layouts and update button functionality

* refactor(build.gradle.kts)

로띠 애니메이션 의존성추가

* 로띠 애니메이션 및 버튼 추가

로띠 애니메이션 및 버튼 추가

* refactor(KaKaoLoginUtil.kt)

로그인 안되서 테스트용 코드 넣음

* refactor(MapActivity.kt)

NavigatableActivity() 삭제 및 로띠 버튼 기능 추가

* refactor(MypageActivity.kt)

네비바 삭제

* delete(NavigatableActivity.kt)

네비바 삭제

* refactor(NoNetworkActivity.kt)

네비바 삭제

* refactor( AnimatedNavButtonHelper.kt)

애니메이션 버튼 관리 오브젝트

* refactor(ic_profile.xml)

중복 파일 삭제

* refactor(activity_map.xml)

네비바 제거, 애니메이션 버튼 추가

* Revert "refactor(KaKaoLoginUtil.kt)"
테스트용 코드라 커밋 revert

This reverts commit a113e9202a53957ef8e9230075307d64975a618d.

* fix activity_map : 지도가 안 보이는 문제 해결

* feature dudoji 앱 로고 변경

* chore KakaoLoginUtil

* refacotr(activity_map.xml)

상점 버튼, 소셜버튼, 마이핀 버튼 추기

* feature(버튼 애니메이션, 아이콘 이미지)

버튼 로띠 애니메이션, 아이콘 이미지 추가

* refactor(AnimatedNavButtonHelper.kt)

프로필만 따로 하던 함수 지우고, 4개 버튼 1개 함수로 통합

* refactor(MapActivity.kt)

oncreate에 넣었던 기능들 setupAnimatedNavButtons() 함수로 생성
icon은 로띠 애니메이션화가 안되서 버튼 lottie애니 20%쯤에 등장하게 구현

* feature(filter_bar_background.json)

필터 바 애니메이션

* refactor(MapActivity.kt)

setupAnimationIcon 별도의 오브젝트 클래스(LottieIconSyncHelper)로 선언

* feature(필터 이미지)

필터 이미지 추가

* refactor(PinFilter.kt)

버튼 클릭에 따라 이미지 변경(필터 꺼지고 켜지는 이미지)

* refactor(LottileIcomSyncHelper.kt)

로띠애니메이션과 아이콘을 비슷한 시간에 생성하기 위해 생성한 클래스

* refactor(activity_map.xml)

필터 버튼 이미지 변경

* refactor(MapActivity.kt)

setFriendFilterButton() 기능setupAnimatedNavButtons()의 social 버튼으로 변경
필터 버튼 및 핀 버튼도 센터 버튼으로 사라졌다가 나타나게 구현

* refactor(PinFilter.kt)

생성자로 버튼을 넘겨줄 필요가 없어서 내부에 선언해줌

* refactor(AnimatedNavButtonHelper.kt)

필터 버튼들과, 핀 버튼도 사라졌다가 나타나므로 해당 오브젝트에 선언해준다.

* refactor(activity_map.xml)

레이아웃 위치 조절 및 filter bar 추가,
로띠 애니메이션으로 적용함

---------

+++ app/src/main/java/com/dudoji/android/follow/FollowAdapter.kt
@@ -24,21 +24,21 @@ class FollowAdapter(private val followings: List<User>, private val activity: Ma
     inner class FollowingViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val name = view.findViewById<TextView>(R.id.textName)
         val desc = view.findViewById<TextView>(R.id.textEmail)
-        val image: ImageView = view.findViewById(R.id.imgProfile)
+        val image: ImageView = view.findViewById(R.id.following_item_image)
         val deleteButton = view.findViewById<Button>(R.id.buttonDelete)
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FollowingViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.item_deletable_following, parent, false)
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.deletable_following_item, parent, false)
         return FollowingViewHolder(view)
     }
 
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onBindViewHolder(holder: FollowingViewHolder, position: Int) {
         val following = followings[position]
-
         holder.name.text = following.name
         holder.desc.text = following.email
+        var isFollowing = true
         val imageUrl = following.profileImageUrl
         if (!imageUrl.isNullOrBlank()) {
             Glide.with(activity)
@@ -49,10 +49,21 @@ class FollowAdapter(private val followings: List<User>, private val activity: Ma
         }
         holder.deleteButton.setOnClickListener {
             activity.lifecycleScope.launch {
-                if (FollowRepository.deleteFollowing(following))
-                    Toast.makeText(activity, "${following.name}를 언팔로우 합니다.", Toast.LENGTH_SHORT).show()
-                else
-                    Toast.makeText(activity, "언팔로우에 실패했습니다.", Toast.LENGTH_SHORT).show()
+                if (!isFollowing) {
+                    FollowRepository.addFollowing(following)
+                    Toast.makeText(activity, "${following.name}를 팔로우 합니다.", Toast.LENGTH_SHORT).show()
+                    holder.deleteButton.text = "팔로우"
+                    holder.deleteButton.setBackgroundColor(activity.getColor(R.color.orange))
+                    isFollowing = true
+                } else {
+                    if (FollowRepository.deleteFollowing(following)){
+                        Toast.makeText(activity, "${following.name}를 언팔로우 합니다.", Toast.LENGTH_SHORT).show()
+                        holder.deleteButton.text = "팔로잉"
+                        holder.deleteButton.setBackgroundColor(activity.getColor(R.color.orange))
+                        isFollowing = false
+                    } else
+                        Toast.makeText(activity, "언팔로우에 실패했습니다.", Toast.LENGTH_SHORT).show()
+                }
             }
         }
     }
